Clustering projection using logistic regression from the LiblineaR R package

The function implements a script that downsamples data a dataset, trains a logistic regression classifier model and then projects its clustering onto itself using a trained L1-regularized logistic regression model.

Usage

LogisticRegression(
  training.sparse.matrix = NULL,
  training.ident = NULL,
  C = 0.3,
  reg.type = "L1",
  test.sparse.matrix = NULL,
  d = 0.3,
  batch.label = NULL,
  training_ident_subset = NULL
)

Arguments

training.sparse.matrix

A sparse matrix (dgCMatrix) containing training sample's feature expression data with features in rows and cells in columns. Default is NULL.

training.ident

A named factor containing sample's cluster labels for each cell in training.sparse.matrix. Default is NULL.

C

Cost of constraints violation in L1-regularized logistic regression (C). Default is 0.3.

reg.type

"L1" for LASSO and "L2" for Ridge. Default is "L1".

test.sparse.matrix

A sparse matrix (dgCMatrix) containing test sample's feature expression data with features in rows and cells in columns. Default is NULL.

d

A numeric smaller than 1 and greater than 0 that determines how many cells per cluster should be down- and oversampled (d in N/k*d), where N is the total number of cells and k the number of clusters. Default is 0.3.

batch.label

A character vector with batch labels corresponding to the cells given in training.ident. The character batch labels need to be named with the cells names given in training.ident. By default NULL, i.e., cells are sampled evenly regardless their batch.

training_ident_subset

A character or numeric vector with cell ids to use as train set. By default NULL. If given, the down- and oversampled parameters are ignored.

Value

a list containing the output of the LiblineaR prediction