Reference mapping

This function allows to project new query data sets onto a reference built with Coralysis as well as transfer cell labels from the reference to queries.

Usage

ReferenceMapping.SingleCellExperiment(
  ref,
  query,
  ref.label,
  scale.query.by,
  project.umap,
  select.icp.models,
  k.nn,
  dimred.name.prefix
)

# S4 method for class 'SingleCellExperiment,SingleCellExperiment'
ReferenceMapping(
  ref,
  query,
  ref.label,
  scale.query.by = NULL,
  project.umap = FALSE,
  select.icp.models = metadata(ref)$coralysis$pca.params$select.icp.tables,
  k.nn = 10,
  dimred.name.prefix = ""
)

Arguments

ref

An object of SingleCellExperiment class trained with Coralysis and after running RunPCA(..., return.model = TRUE) function.

query

An object of SingleCellExperiment class to project onto ref.

ref.label

A character cell metadata column name from the ref object to transfer to the queries.

scale.query.by

Should the query data be scaled by cell or by feature. By default is NULL, i.e., is not scaled. Scale it if reference was scaled.

project.umap

Project query data onto reference UMAP (logical). By default FALSE. If TRUE, the ref object needs to have a UMAP embedding obtained with RunUMAP(..., return.model = TRUE) function.

k.nn

The number of k nearest neighbors to use in the classification KNN algorithm used to transfer labels from the reference to queries (integer). By default 10.

dimred.name.prefix

Dimensional reduction name prefix to add to the computed PCA and UMAP. By default nothing is added, i.e., dimred.name.prefix = "".

selecte.icp.models

Select the reference ICP models to use for query cluster probability prediction. By default metadata(ref)$coralysis$pca.params$select.icp.tables, i.e., the models selected to compute the reference PCA are selected. If NULL all are used. Otherwise a numeric vector should be given to select the ICP models of interest.

Value

An object of SingleCellExperiment class.

Examples

## Reference-mapping
# Split the SCE object by 'batch'
ref <- pbmc_10Xassays[,pbmc_10Xassays$batch=="V2"] # let V2 assay batch be the reference data set
query <- pbmc_10Xassays[,pbmc_10Xassays$batch=="V1"] # let V1 be the query (unknown annotations)

# 1) Train the reference - 'L = 4' just for highlighting purposes; use 'L=50' (default) or greater
set.seed(123)
ref <- RunParallelDivisiveICP(object = ref, L = 4, divisive.method = "cluster", threads = 1) # runs without 'batch.label' as it represents 1 sample only
#> 
#> Building training set...
#> Training set successfully built.
#> 
#> Computing cluster seed.
#> Parallelism disabled, because threads = 1
#> 
#> Initializing divisive ICP clustering...
#> ICP run: 1
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 244 256
#> ARI=0.860905413892533
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 244 256
#> projected clustering = 249 251
#> ARI=0.913763282439273
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 249 251
#> projected clustering = 249 251
#> ARI=0.952480774125568
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 249 251
#> projected clustering = 256 244
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 249 251
#> projected clustering = 253 247
#> ARI=0.952480786241631
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 253 247
#> projected clustering = 254 246
#> ARI=0.960320512330166
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 254 246
#> projected clustering = 253 247
#> ARI=0.976096114700718
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 252 248
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 249 251
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 250 250
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 256 244
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 255 245
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 124 123 127 126
#> projected clustering = 113 128 132 127
#> ARI=0.54068280765863
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 113 128 132 127
#> projected clustering = 113 130 130 127
#> ARI=0.756503050272598
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 113 130 130 127
#> projected clustering = 113 128 132 127
#> ARI=0.851546098899705
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 113 128 132 127
#> projected clustering = 111 136 125 128
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 113 128 132 127
#> projected clustering = 98 139 136 127
#> ARI=0.854439118323419
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 98 139 136 127
#> projected clustering = 99 142 132 127
#> ARI=0.935978277338942
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 99 142 132 127
#> projected clustering = 102 136 135 127
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 99 142 132 127
#> projected clustering = 119 130 122 129
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 99 142 132 127
#> projected clustering = 100 133 140 127
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 99 142 132 127
#> projected clustering = 105 129 139 127
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 99 142 132 127
#> projected clustering = 104 137 131 128
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 50 49 66 66 64 63 71 71
#> projected clustering = 38 68 50 90 44 83 45 82
#> ARI=0.53288324427006
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 38 68 50 90 44 83 45 82
#> projected clustering = 43 76 44 98 42 85 39 73
#> ARI=0.710931948656894
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 43 76 44 98 42 85 39 73
#> projected clustering = 52 69 45 99 42 84 39 70
#> ARI=0.837897960017352
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 52 69 45 99 42 84 39 70
#> projected clustering = 49 73 47 96 48 78 44 65
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 52 69 45 99 42 84 39 70
#> projected clustering = 49 72 47 96 40 87 43 66
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 52 69 45 99 42 84 39 70
#> projected clustering = 47 79 51 97 41 84 35 66
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 52 69 45 99 42 84 39 70
#> projected clustering = 55 69 47 98 45 82 38 66
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 52 69 45 99 42 84 39 70
#> projected clustering = 42 82 48 97 41 86 36 68
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 35 34 23 22 21 21 26 26 50 49 35 35 20 19 42 42
#> projected clustering = 41 40 20 28 18 20 20 19 31 65 28 37 14 30 51 38
#> ARI=0.467479816147333
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 41 40 20 28 18 20 20 19 31 65 28 37 14 30 51 38
#> projected clustering = 29 61 19 30 28 19 18 19 33 61 26 38 14 25 38 42
#> ARI=0.574805421986085
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 29 61 19 30 28 19 18 19 33 61 26 38 14 25 38 42
#> projected clustering = 28 61 22 27 21 25 27 20 35 55 19 36 16 27 41 40
#> ARI=0.656355858516036
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 28 61 22 27 21 25 27 20 35 55 19 36 16 27 41 40
#> projected clustering = 28 61 18 33 36 23 23 15 46 45 25 35 15 31 37 29
#> ARI=0.695996650729343
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 28 61 18 33 36 23 23 15 46 45 25 35 15 31 37 29
#> projected clustering = 32 51 17 33 29 26 24 18 49 44 22 36 17 32 42 28
#> ARI=0.737841514204607
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 32 51 17 33 29 26 24 18 49 44 22 36 17 32 42 28
#> projected clustering = 27 48 20 29 26 35 29 18 50 44 20 36 20 34 42 22
#> ARI=0.746191345037651
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 27 48 20 29 26 35 29 18 50 44 20 36 20 34 42 22
#> projected clustering = 24 59 22 27 25 36 21 20 53 41 19 34 22 31 44 22
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 27 48 20 29 26 35 29 18 50 44 20 36 20 34 42 22
#> projected clustering = 44 38 22 27 32 32 26 14 55 38 19 35 22 34 45 17
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 27 48 20 29 26 35 29 18 50 44 20 36 20 34 42 22
#> projected clustering = 35 43 22 26 28 36 28 16 49 44 19 35 20 37 41 21
#> ARI=0.782069010822919
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 35 43 22 26 28 36 28 16 49 44 19 35 20 37 41 21
#> projected clustering = 35 42 25 25 25 43 31 13 42 50 18 36 18 38 36 23
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 35 43 22 26 28 36 28 16 49 44 19 35 20 37 41 21
#> projected clustering = 37 40 19 29 26 37 30 15 49 44 19 35 23 34 40 23
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 35 43 22 26 28 36 28 16 49 44 19 35 20 37 41 21
#> projected clustering = 31 51 21 27 28 27 16 26 45 48 22 29 25 32 31 41
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 35 43 22 26 28 36 28 16 49 44 19 35 20 37 41 21
#> projected clustering = 26 52 24 21 44 27 40 16 43 50 19 32 23 27 37 19
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 35 43 22 26 28 36 28 16 49 44 19 35 20 37 41 21
#> projected clustering = 33 47 20 29 26 41 29 17 42 51 17 37 21 32 31 27
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.782069010822919.
#> ICP run: 2
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 251 249
#> ARI=0.868360197000476
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 251 249
#> projected clustering = 245 255
#> ARI=0.936897594475247
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 245 255
#> projected clustering = 253 247
#> ARI=0.936897607450507
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 253 247
#> projected clustering = 254 246
#> ARI=0.976096114700718
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 254 246
#> projected clustering = 254 246
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 254 246
#> projected clustering = 253 247
#> ARI=0.991999974001112
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 253 247
#> projected clustering = 253 247
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 253 247
#> projected clustering = 248 252
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 253 247
#> projected clustering = 253 247
#> ARI=1
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 249 251
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 258 242
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 252 248
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 255 245
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 253 247
#> projected clustering = 254 246
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 124 123 127 126
#> projected clustering = 122 115 135 128
#> ARI=0.553659855090899
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 122 115 135 128
#> projected clustering = 102 133 137 128
#> ARI=0.725803255330205
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 102 133 137 128
#> projected clustering = 107 125 141 127
#> ARI=0.883923474090111
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 107 125 141 127
#> projected clustering = 125 106 142 127
#> ARI=0.900710010389371
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 125 106 142 127
#> projected clustering = 129 107 136 128
#> ARI=0.930836807718945
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 129 107 136 128
#> projected clustering = 132 112 128 128
#> ARI=0.94400292843619
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 132 112 128 128
#> projected clustering = 131 109 132 128
#> ARI=0.968998872881955
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 131 109 132 128
#> projected clustering = 132 107 134 127
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 131 109 132 128
#> projected clustering = 134 116 122 128
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 131 109 132 128
#> projected clustering = 130 106 137 127
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 131 109 132 128
#> projected clustering = 132 110 131 127
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 131 109 132 128
#> projected clustering = 132 116 125 127
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 66 65 66 66 64 64 55 54
#> projected clustering = 44 84 41 95 51 78 57 50
#> ARI=0.554148391024381
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 44 84 41 95 51 78 57 50
#> projected clustering = 56 75 46 97 36 91 49 50
#> ARI=0.735349930931027
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 56 75 46 97 36 91 49 50
#> projected clustering = 53 78 45 98 27 101 48 50
#> ARI=0.795744827023885
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 53 78 45 98 27 101 48 50
#> projected clustering = 33 99 46 96 23 106 48 49
#> ARI=0.86127283933337
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 33 99 46 96 23 106 48 49
#> projected clustering = 34 96 46 98 25 104 55 42
#> ARI=0.910390119908751
#> probability matrix dimensions = 500 8
#> EPOCH: 6
#> current clustering = 34 96 46 98 25 104 55 42
#> projected clustering = 26 105 46 96 25 105 58 39
#> ARI=0.920639888328548
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 26 105 46 96 25 105 58 39
#> projected clustering = 35 95 45 99 29 100 53 44
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 26 105 46 96 25 105 58 39
#> projected clustering = 31 102 46 98 32 99 53 39
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 26 105 46 96 25 105 58 39
#> projected clustering = 38 93 45 97 24 107 55 41
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 26 105 46 96 25 105 58 39
#> projected clustering = 28 104 45 97 26 104 57 39
#> ARI=0.945156486477193
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 28 104 45 97 26 104 57 39
#> projected clustering = 36 94 44 97 35 96 56 42
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 28 104 45 97 26 104 57 39
#> projected clustering = 41 92 45 98 28 102 54 40
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 28 104 45 97 26 104 57 39
#> projected clustering = 33 98 46 98 32 98 54 41
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 28 104 45 97 26 104 57 39
#> projected clustering = 46 83 45 97 26 103 61 39
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 28 104 45 97 26 104 57 39
#> projected clustering = 35 97 45 97 29 101 57 39
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 52 52 23 22 52 52 49 48 29 28 20 19 14 14 13 13
#> projected clustering = 43 48 18 26 26 76 36 57 18 38 28 17 28 13 12 16
#> ARI=0.482638813087485
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 43 48 18 26 26 76 36 57 18 38 28 17 28 13 12 16
#> projected clustering = 31 61 21 23 23 79 42 51 22 29 23 22 31 14 13 15
#> ARI=0.637387065359217
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 31 61 21 23 23 79 42 51 22 29 23 22 31 14 13 15
#> projected clustering = 27 58 21 25 25 76 35 59 29 21 25 22 29 18 12 18
#> ARI=0.687694938631591
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 27 58 21 25 25 76 35 59 29 21 25 22 29 18 12 18
#> projected clustering = 40 47 16 30 38 64 25 70 26 27 20 27 27 15 11 17
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 27 58 21 25 25 76 35 59 29 21 25 22 29 18 12 18
#> projected clustering = 30 52 21 25 36 65 40 53 21 24 31 22 33 18 11 18
#> ARI=0.691849059657159
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 30 52 21 25 36 65 40 53 21 24 31 22 33 18 11 18
#> projected clustering = 35 46 17 29 39 52 38 55 20 26 27 29 31 17 11 28
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 30 52 21 25 36 65 40 53 21 24 31 22 33 18 11 18
#> projected clustering = 21 53 18 29 30 69 34 59 17 26 30 25 36 21 11 21
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 30 52 21 25 36 65 40 53 21 24 31 22 33 18 11 18
#> projected clustering = 31 50 14 30 44 57 43 51 25 25 25 25 29 21 11 19
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 30 52 21 25 36 65 40 53 21 24 31 22 33 18 11 18
#> projected clustering = 36 50 16 30 41 57 46 47 23 28 23 26 29 15 12 21
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 30 52 21 25 36 65 40 53 21 24 31 22 33 18 11 18
#> projected clustering = 40 45 22 25 39 60 40 52 21 24 24 29 28 21 11 19
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.691849059657159.
#> ICP run: 3
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 250 250
#> ARI=0.84609156626506
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 250 250
#> projected clustering = 262 238
#> ARI=0.890917902259344
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 262 238
#> projected clustering = 259 241
#> ARI=0.97609619752109
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 259 241
#> projected clustering = 263 237
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 259 241
#> projected clustering = 255 245
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 259 241
#> projected clustering = 261 239
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 259 241
#> projected clustering = 255 245
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 259 241
#> projected clustering = 258 242
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 121 120 130 129
#> projected clustering = 96 147 118 139
#> ARI=0.500341997551534
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 96 147 118 139
#> projected clustering = 96 148 127 129
#> ARI=0.664052960506769
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 96 148 127 129
#> projected clustering = 113 134 126 127
#> ARI=0.688633924701096
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 113 134 126 127
#> projected clustering = 113 140 118 129
#> ARI=0.8678779191565
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 113 140 118 129
#> projected clustering = 116 146 110 128
#> ARI=0.876202678459898
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 116 146 110 128
#> projected clustering = 129 142 100 129
#> ARI=0.899252160617905
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 129 142 100 129
#> projected clustering = 123 147 99 131
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 129 142 100 129
#> projected clustering = 122 149 99 130
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 129 142 100 129
#> projected clustering = 132 140 99 129
#> ARI=0.908859915523874
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 132 140 99 129
#> projected clustering = 136 137 99 128
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 132 140 99 129
#> projected clustering = 126 145 100 129
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 132 140 99 129
#> projected clustering = 129 144 97 130
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 132 140 99 129
#> projected clustering = 122 150 100 128
#> ARI=0.918579235817529
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 122 150 100 128
#> projected clustering = 123 147 100 130
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 122 150 100 128
#> projected clustering = 124 149 97 130
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 122 150 100 128
#> projected clustering = 130 144 97 129
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 122 150 100 128
#> projected clustering = 123 148 98 131
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 122 150 100 128
#> projected clustering = 123 148 101 128
#> ARI=0.940005413311393
#> probability matrix dimensions = 500 4
#> EPOCH: 10
#> current clustering = 123 148 101 128
#> projected clustering = 122 150 100 128
#> ARI=0.961621265809584
#> probability matrix dimensions = 500 4
#> EPOCH: 11
#> current clustering = 122 150 100 128
#> projected clustering = 124 149 99 128
#> ARI=0.967028830914731
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 124 149 99 128
#> projected clustering = 131 144 97 128
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 124 149 99 128
#> projected clustering = 122 153 95 130
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 124 149 99 128
#> projected clustering = 125 148 99 128
#> ARI=0.978648486140703
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 125 148 99 128
#> projected clustering = 130 145 98 127
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 125 148 99 128
#> projected clustering = 123 150 99 128
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 125 148 99 128
#> projected clustering = 125 148 99 128
#> ARI=1
#> probability matrix dimensions = 500 4
#> EPOCH: 14
#> current clustering = 125 148 99 128
#> projected clustering = 125 147 100 128
#> probability matrix dimensions = 500 4
#> EPOCH: 14
#> current clustering = 125 148 99 128
#> projected clustering = 129 146 98 127
#> probability matrix dimensions = 500 4
#> EPOCH: 14
#> current clustering = 125 148 99 128
#> projected clustering = 123 150 99 128
#> probability matrix dimensions = 500 4
#> EPOCH: 14
#> current clustering = 125 148 99 128
#> projected clustering = 125 148 99 128
#> probability matrix dimensions = 500 4
#> EPOCH: 14
#> current clustering = 125 148 99 128
#> projected clustering = 128 145 99 128
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 74 74 63 62 64 64 50 49
#> projected clustering = 36 110 55 70 53 75 43 58
#> ARI=0.572407431101838
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 36 110 55 70 53 75 43 58
#> projected clustering = 41 104 67 62 45 82 49 50
#> ARI=0.742361262727057
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 41 104 67 62 45 82 49 50
#> projected clustering = 32 113 70 60 45 80 43 57
#> ARI=0.792346930293855
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 32 113 70 60 45 80 43 57
#> projected clustering = 33 115 67 62 43 82 41 57
#> ARI=0.830107233044497
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 33 115 67 62 43 82 41 57
#> projected clustering = 34 115 63 64 36 89 40 59
#> ARI=0.846952612144897
#> probability matrix dimensions = 500 8
#> EPOCH: 6
#> current clustering = 34 115 63 64 36 89 40 59
#> projected clustering = 34 115 65 61 28 98 37 62
#> ARI=0.879422774081456
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 34 115 65 61 28 98 37 62
#> projected clustering = 40 108 68 60 48 77 37 62
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 34 115 65 61 28 98 37 62
#> projected clustering = 40 111 62 62 40 85 34 66
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 34 115 65 61 28 98 37 62
#> projected clustering = 31 120 67 57 38 89 38 60
#> ARI=0.879683026407296
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 31 120 67 57 38 89 38 60
#> projected clustering = 44 107 65 57 38 90 31 68
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 31 120 67 57 38 89 38 60
#> projected clustering = 39 112 63 62 33 93 38 60
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 31 120 67 57 38 89 38 60
#> projected clustering = 34 117 72 54 41 85 34 63
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 31 120 67 57 38 89 38 60
#> projected clustering = 33 123 66 53 41 85 33 66
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 31 120 67 57 38 89 38 60
#> projected clustering = 28 123 62 59 45 81 40 62
#> ARI=0.883566499104225
#> probability matrix dimensions = 500 8
#> EPOCH: 9
#> current clustering = 28 123 62 59 45 81 40 62
#> projected clustering = 34 122 63 56 45 80 38 62
#> ARI=0.896074481290736
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 34 122 63 56 45 80 38 62
#> projected clustering = 38 121 66 50 47 78 40 60
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 34 122 63 56 45 80 38 62
#> projected clustering = 38 115 60 62 45 81 33 66
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 34 122 63 56 45 80 38 62
#> projected clustering = 29 125 68 53 48 77 38 62
#> ARI=0.942576336509811
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 29 125 68 53 48 77 38 62
#> projected clustering = 32 119 66 56 52 74 37 64
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 29 125 68 53 48 77 38 62
#> projected clustering = 27 125 77 45 48 77 39 62
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 29 125 68 53 48 77 38 62
#> projected clustering = 31 126 61 56 46 79 37 64
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 29 125 68 53 48 77 38 62
#> projected clustering = 32 125 63 56 47 77 35 65
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 29 125 68 53 48 77 38 62
#> projected clustering = 36 121 65 50 53 74 39 62
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 63 62 27 26 39 38 31 31 19 19 15 14 34 34 24 24
#> projected clustering = 47 74 15 37 28 43 30 30 18 23 16 16 18 52 23 30
#> ARI=0.45948713735848
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 47 74 15 37 28 43 30 30 18 23 16 16 18 52 23 30
#> projected clustering = 21 90 15 42 34 39 37 26 21 18 21 23 13 48 22 30
#> ARI=0.591166893295404
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 21 90 15 42 34 39 37 26 21 18 21 23 13 48 22 30
#> projected clustering = 30 72 21 37 34 24 34 25 17 25 30 23 14 47 32 35
#> ARI=0.605981034542905
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 30 72 21 37 34 24 34 25 17 25 30 23 14 47 32 35
#> projected clustering = 28 78 18 41 43 23 24 30 23 24 25 20 13 48 31 31
#> ARI=0.685320634494923
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 28 78 18 41 43 23 24 30 23 24 25 20 13 48 31 31
#> projected clustering = 26 77 19 39 49 15 25 30 22 24 25 17 18 52 38 24
#> ARI=0.709636350612721
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 26 77 19 39 49 15 25 30 22 24 25 17 18 52 38 24
#> projected clustering = 25 77 20 34 42 20 19 33 27 24 24 22 15 54 34 30
#> ARI=0.761841683176744
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 25 77 20 34 42 20 19 33 27 24 24 22 15 54 34 30
#> projected clustering = 22 81 26 27 43 24 17 25 32 28 25 19 13 58 33 27
#> ARI=0.771289006286334
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 22 81 26 27 43 24 17 25 32 28 25 19 13 58 33 27
#> projected clustering = 30 75 26 25 43 18 18 26 34 25 24 21 12 56 41 26
#> ARI=0.792424241130767
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 30 75 26 25 43 18 18 26 34 25 24 21 12 56 41 26
#> projected clustering = 31 71 28 30 39 23 19 32 32 22 24 18 16 49 36 30
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 30 75 26 25 43 18 18 26 34 25 24 21 12 56 41 26
#> projected clustering = 36 68 22 28 45 15 22 26 33 22 26 19 15 56 42 25
#> ARI=0.811046532390255
#> probability matrix dimensions = 500 16
#> EPOCH: 10
#> current clustering = 36 68 22 28 45 15 22 26 33 22 26 19 15 56 42 25
#> projected clustering = 33 71 29 25 51 15 26 30 32 15 23 23 14 51 37 25
#> probability matrix dimensions = 500 16
#> EPOCH: 10
#> current clustering = 36 68 22 28 45 15 22 26 33 22 26 19 15 56 42 25
#> projected clustering = 32 71 26 26 44 14 21 34 29 19 24 19 19 53 48 21
#> probability matrix dimensions = 500 16
#> EPOCH: 10
#> current clustering = 36 68 22 28 45 15 22 26 33 22 26 19 15 56 42 25
#> projected clustering = 32 68 22 28 47 20 26 28 31 17 26 19 14 62 32 28
#> probability matrix dimensions = 500 16
#> EPOCH: 10
#> current clustering = 36 68 22 28 45 15 22 26 33 22 26 19 15 56 42 25
#> projected clustering = 30 73 25 29 42 16 22 27 33 20 23 23 15 53 39 30
#> ARI=0.830918787052101
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 30 73 25 29 42 16 22 27 33 20 23 23 15 53 39 30
#> projected clustering = 41 64 24 29 35 29 22 23 34 21 30 14 18 51 38 27
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 30 73 25 29 42 16 22 27 33 20 23 23 15 53 39 30
#> projected clustering = 34 68 27 28 45 17 18 31 32 20 28 23 20 43 33 33
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 30 73 25 29 42 16 22 27 33 20 23 23 15 53 39 30
#> projected clustering = 33 69 25 31 41 22 21 29 35 18 24 21 18 46 38 29
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 30 73 25 29 42 16 22 27 33 20 23 23 15 53 39 30
#> projected clustering = 37 72 31 28 42 17 25 24 33 19 21 23 14 44 33 37
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 30 73 25 29 42 16 22 27 33 20 23 23 15 53 39 30
#> projected clustering = 33 64 24 35 38 24 21 31 28 23 31 20 13 52 34 29
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.830918787052101.
#> ICP run: 4
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 243 257
#> ARI=0.838733073159178
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 243 257
#> projected clustering = 247 253
#> ARI=0.906116019728455
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 247 253
#> projected clustering = 234 266
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 247 253
#> projected clustering = 237 263
#> ARI=0.906116355627711
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 237 263
#> projected clustering = 230 270
#> ARI=0.913763258855457
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 230 270
#> projected clustering = 229 271
#> ARI=0.929152726197092
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 229 271
#> projected clustering = 229 271
#> ARI=1
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 229 271
#> projected clustering = 229 271
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 229 271
#> projected clustering = 229 271
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 229 271
#> projected clustering = 228 272
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 229 271
#> projected clustering = 226 274
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 229 271
#> projected clustering = 227 273
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 136 135 115 114
#> projected clustering = 125 146 110 119
#> ARI=0.547906298581332
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 125 146 110 119
#> projected clustering = 139 134 96 131
#> ARI=0.549437446273372
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 139 134 96 131
#> projected clustering = 141 128 100 131
#> ARI=0.698123810252216
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 141 128 100 131
#> projected clustering = 139 131 101 129
#> ARI=0.884265998470723
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 139 131 101 129
#> projected clustering = 142 128 101 129
#> ARI=0.971821294900963
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 142 128 101 129
#> projected clustering = 142 131 100 127
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 142 128 101 129
#> projected clustering = 143 126 103 128
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 142 128 101 129
#> projected clustering = 141 130 98 131
#> ARI=0.97977753803617
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 141 130 98 131
#> projected clustering = 140 133 97 130
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 141 130 98 131
#> projected clustering = 141 134 96 129
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 141 130 98 131
#> projected clustering = 148 123 99 130
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 141 130 98 131
#> projected clustering = 139 134 100 127
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 141 130 98 131
#> projected clustering = 136 134 102 128
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 71 70 65 65 66 65 49 49
#> projected clustering = 43 102 61 68 58 71 47 50
#> ARI=0.503454161655723
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 43 102 61 68 58 71 47 50
#> projected clustering = 42 106 60 68 57 70 40 57
#> ARI=0.727553082864517
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 42 106 60 68 57 70 40 57
#> projected clustering = 39 104 72 59 64 63 45 54
#> ARI=0.734959448295588
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 39 104 72 59 64 63 45 54
#> projected clustering = 56 88 71 59 65 62 43 56
#> ARI=0.77721746147178
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 56 88 71 59 65 62 43 56
#> projected clustering = 50 93 73 59 57 69 43 56
#> ARI=0.822698616399041
#> probability matrix dimensions = 500 8
#> EPOCH: 6
#> current clustering = 50 93 73 59 57 69 43 56
#> projected clustering = 54 93 68 61 71 55 44 54
#> ARI=0.876640439779276
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 54 93 68 61 71 55 44 54
#> projected clustering = 50 98 68 61 72 54 43 54
#> ARI=0.88138811235916
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 50 98 68 61 72 54 43 54
#> projected clustering = 44 100 69 63 69 58 43 54
#> ARI=0.895793640431056
#> probability matrix dimensions = 500 8
#> EPOCH: 9
#> current clustering = 44 100 69 63 69 58 43 54
#> projected clustering = 44 103 65 64 75 51 44 54
#> ARI=0.913112993361468
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 44 103 65 64 75 51 44 54
#> projected clustering = 46 100 68 64 68 58 41 55
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 44 103 65 64 75 51 44 54
#> projected clustering = 46 100 70 60 70 56 44 54
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 44 103 65 64 75 51 44 54
#> projected clustering = 48 95 65 66 79 48 40 59
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 44 103 65 64 75 51 44 54
#> projected clustering = 48 97 66 63 74 54 41 57
#> ARI=0.930706563065716
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 48 97 66 63 74 54 41 57
#> projected clustering = 49 98 62 63 81 48 42 57
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 48 97 66 63 74 54 41 57
#> projected clustering = 48 96 68 63 68 58 41 58
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 48 97 66 63 74 54 41 57
#> projected clustering = 50 99 65 61 78 49 38 60
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 48 97 66 63 74 54 41 57
#> projected clustering = 47 97 69 61 75 53 44 54
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 48 97 66 63 74 54 41 57
#> projected clustering = 47 98 67 63 82 45 42 56
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 49 48 32 31 27 27 24 24 21 20 33 33 29 28 37 37
#> projected clustering = 18 80 24 38 30 29 28 22 25 20 18 47 21 33 28 39
#> ARI=0.562965303051387
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 18 80 24 38 30 29 28 22 25 20 18 47 21 33 28 39
#> projected clustering = 15 81 23 38 34 23 28 28 26 22 29 32 21 31 33 36
#> ARI=0.754901730972879
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 15 81 23 38 34 23 28 28 26 22 29 32 21 31 33 36
#> projected clustering = 37 60 19 43 34 26 30 24 28 24 26 36 19 27 26 41
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 15 81 23 38 34 23 28 28 26 22 29 32 21 31 33 36
#> projected clustering = 22 73 25 37 34 26 29 29 23 24 25 35 20 30 31 37
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 15 81 23 38 34 23 28 28 26 22 29 32 21 31 33 36
#> projected clustering = 18 76 26 41 41 20 33 26 24 23 28 28 26 24 31 35
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 15 81 23 38 34 23 28 28 26 22 29 32 21 31 33 36
#> projected clustering = 25 61 25 34 35 15 36 30 26 27 17 46 14 32 44 33
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 15 81 23 38 34 23 28 28 26 22 29 32 21 31 33 36
#> projected clustering = 17 75 23 33 33 24 24 36 24 24 39 26 28 25 32 37
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.754901730972879.
#> 
#> Divisive ICP clustering completed successfully.
#> 
#> Predicting cell cluster probabilities using ICP models...
#> Prediction of cell cluster probabilities completed successfully.
#> 
#> Multi-level integration completed successfully.

# 2) Compute reference PCA & UMAP
ref <- RunPCA(ref, return.model = TRUE, pca.method = "stats")
#> Divisive ICP: selecting ICP tables multiple of 4
set.seed(123)
ref <- RunUMAP(ref, return.model = TRUE)

# 3) Project & predict query cell labels 
set.seed(1024)
map <- ReferenceMapping(ref = ref, query = query, ref.label = "cell_type", scale.query.by = NULL, project.umap = TRUE)

# Calculate accuracy
preds_x_truth <- table(map$coral_labels, map$cell_type)
preds_x_truth
#>                 
#>                  aDC B mem B naive CD4 mem CD4 naive CD8 eff CD8 T HSC
#>   aDC             14     0       0       0         0       0     0   0
#>   B mem            0    35      11       0         0       1     0   1
#>   B naive          0     8      74       0         0       0     0   0
#>   CD4 mem          0     0       0      54        20       3     1   0
#>   CD4 naive        0     1       0      69       169       0    18   4
#>   CD8 eff          0     1       0       2         1      96     1   0
#>   CD8 T            0     0       0       3         6       8    62   0
#>   HSC              0     0       0       0         0       0     0   0
#>   Megakaryocyte    0     0       0       0         0       0     0   0
#>   Monocyte         1     0       0       0         0       0     0   0
#>   CD16+ monocyte   0     0       0       0         0       0     0   0
#>   NK               0     0       0       0         0       0     0   0
#>   pDC              0     0       0       0         0       0     0   0
#>   Treg             0     0       0       0         0       0     0   0
#>                 
#>                  Megakaryocyte Monocyte CD16+ monocyte  NK pDC Treg
#>   aDC                        0        0              0   0   0    0
#>   B mem                      0        0              0   0   0    0
#>   B naive                    0        0              0   0   0    0
#>   CD4 mem                    0        0              0   0   0    3
#>   CD4 naive                  0        0              0   0   0   12
#>   CD8 eff                    1        0              0   2   0    0
#>   CD8 T                      1        0              0   0   0    0
#>   HSC                        0        0              0   0   0    0
#>   Megakaryocyte              0        0              0   0   0    0
#>   Monocyte                   3      177             18   1   0    0
#>   CD16+ monocyte             1        1             58   0   0    0
#>   NK                         0        0              0  54   0    0
#>   pDC                        0        0              0   0   4    0
#>   Treg                       0        0              0   0   0    0
sum(diag(preds_x_truth)) / sum(preds_x_truth) # 0.806 : 81%
#> [1] 0.797