Reference mapping

This function allows to project new query data sets onto a reference built with Coralysis as well as transfer cell labels from the reference to queries.

Usage

ReferenceMapping.SingleCellExperiment(
  ref,
  query,
  ref.label,
  scale.query.by,
  project.umap,
  select.icp.models,
  k.nn,
  dimred.name.prefix
)

# S4 method for SingleCellExperiment,SingleCellExperiment
ReferenceMapping(
  ref,
  query,
  ref.label,
  scale.query.by = NULL,
  project.umap = FALSE,
  select.icp.models = metadata(ref)$coralysis$pca.params$select.icp.tables,
  k.nn = 10,
  dimred.name.prefix = ""
)

Arguments

ref

An object of SingleCellExperiment class trained with Coralysis and after running RunPCA(..., return.model = TRUE) function.

query

An object of SingleCellExperiment class to project onto ref.

ref.label

A character cell metadata column name from the ref object to transfer to the queries.

scale.query.by

Should the query data be scaled by cell or by feature. By default is NULL, i.e., is not scaled. Scale it if reference was scaled.

project.umap

Project query data onto reference UMAP (logical). By default FALSE. If TRUE, the ref object needs to have a UMAP embedding obtained with RunUMAP(..., return.model = TRUE) function.

k.nn

The number of k nearest neighbors to use in the classification KNN algorithm used to transfer labels from the reference to queries (integer). By default 10.

dimred.name.prefix

Dimensional reduction name prefix to add to the computed PCA and UMAP. By default nothing is added, i.e., dimred.name.prefix = "".

selecte.icp.models

Select the reference ICP models to use for query cluster probability prediction. By default metadata(ref)$coralysis$pca.params$select.icp.tables, i.e., the models selected to compute the reference PCA are selected. If NULL all are used. Otherwise a numeric vector should be given to select the ICP models of interest.

Value

An object of SingleCellExperiment class.

Examples

## Reference-mapping
# Split the SCE object by 'batch'
ref <- pbmc_10Xassays[,pbmc_10Xassays$batch=="V2"] # let V2 assay batch be the reference data set
query <- pbmc_10Xassays[,pbmc_10Xassays$batch=="V1"] # let V1 be the query (unknown annotations)

# 1) Train the reference - 'L = 4' just for highlighting purposes; use 'L=50' (default) or greater
set.seed(123)
ref <- RunParallelDivisiveICP(object = ref, L = 4, divisive.method = "cluster", threads = 1) # runs without 'batch.label' as it represents 1 sample only
#> 
#> Building training set...
#> Training set successfully built.
#> 
#> Computing cluster seed.
#> Warning: useNames = NA is deprecated. Instead, specify either useNames = TRUE or useNames = FALSE.
#> Parallelism disabled, because threads = 1
#> 
#> Initializing divisive ICP clustering...
#> ICP run: 1
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 243 257
#> ARI=0.824111794972596
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 243 257
#> projected clustering = 253 247
#> ARI=0.890917588495972
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 253 247
#> projected clustering = 255 245
#> ARI=0.906115952548316
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 255 245
#> projected clustering = 254 246
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 255 245
#> projected clustering = 255 245
#> ARI=0.921442683857967
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 255 245
#> projected clustering = 254 246
#> ARI=0.960320530116754
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 254 246
#> projected clustering = 269 231
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 254 246
#> projected clustering = 252 248
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 254 246
#> projected clustering = 260 240
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 254 246
#> projected clustering = 250 250
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 254 246
#> projected clustering = 251 249
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 123 123 127 127
#> projected clustering = 104 134 129 133
#> ARI=0.520819635224991
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 104 134 129 133
#> projected clustering = 111 127 134 128
#> ARI=0.670161331317624
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 111 127 134 128
#> projected clustering = 119 117 139 125
#> ARI=0.781665569263852
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 119 117 139 125
#> projected clustering = 119 114 124 143
#> ARI=0.820318265538334
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 119 114 124 143
#> projected clustering = 117 111 114 158
#> ARI=0.835943289259642
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 117 111 114 158
#> projected clustering = 120 105 119 156
#> ARI=0.854033281567644
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 120 105 119 156
#> projected clustering = 123 102 122 153
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 120 105 119 156
#> projected clustering = 125 101 127 147
#> ARI=0.857939738942163
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 125 101 127 147
#> projected clustering = 123 100 138 139
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 125 101 127 147
#> projected clustering = 128 97 126 149
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 125 101 127 147
#> projected clustering = 127 98 123 152
#> ARI=0.862546483726132
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 127 98 123 152
#> projected clustering = 126 97 130 147
#> ARI=0.89636150946619
#> probability matrix dimensions = 500 4
#> EPOCH: 10
#> current clustering = 126 97 130 147
#> projected clustering = 126 97 139 138
#> ARI=0.938010963844231
#> probability matrix dimensions = 500 4
#> EPOCH: 11
#> current clustering = 126 97 139 138
#> projected clustering = 126 97 142 135
#> ARI=0.982574507597867
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 126 97 142 135
#> projected clustering = 126 96 145 133
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 126 97 142 135
#> projected clustering = 129 94 143 134
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 126 97 142 135
#> projected clustering = 126 97 138 139
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 126 97 142 135
#> projected clustering = 127 96 145 132
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 126 97 142 135
#> projected clustering = 126 97 138 139
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 49 48 68 67 63 63 71 71
#> projected clustering = 38 59 52 88 47 79 72 65
#> ARI=0.494301551803411
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 38 59 52 88 47 79 72 65
#> projected clustering = 35 64 49 89 47 79 63 74
#> ARI=0.707363021678207
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 35 64 49 89 47 79 63 74
#> projected clustering = 44 55 56 79 42 85 63 76
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 35 64 49 89 47 79 63 74
#> projected clustering = 31 68 50 93 44 82 75 57
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 35 64 49 89 47 79 63 74
#> projected clustering = 34 64 69 76 46 80 59 72
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 35 64 49 89 47 79 63 74
#> projected clustering = 33 63 69 70 48 79 66 72
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 35 64 49 89 47 79 63 74
#> projected clustering = 31 68 69 71 38 87 69 67
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 32 32 25 24 40 39 45 44 32 31 37 37 24 23 18 17
#> projected clustering = 19 39 35 15 38 32 41 49 39 37 15 46 32 23 26 14
#> ARI=0.350545257443073
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 19 39 35 15 38 32 41 49 39 37 15 46 32 23 26 14
#> projected clustering = 24 34 40 31 36 32 32 40 35 29 25 44 32 25 21 20
#> ARI=0.477587994405075
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 24 34 40 31 36 32 32 40 35 29 25 44 32 25 21 20
#> projected clustering = 23 31 29 37 30 43 25 49 47 29 27 32 25 29 25 19
#> ARI=0.496092643178161
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 23 31 29 37 30 43 25 49 47 29 27 32 25 29 25 19
#> projected clustering = 23 32 37 28 37 43 30 47 42 31 22 39 19 28 19 23
#> ARI=0.62298932353608
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 23 32 37 28 37 43 30 47 42 31 22 39 19 28 19 23
#> projected clustering = 30 22 32 27 40 36 35 44 43 42 20 33 25 26 22 23
#> ARI=0.635287298523067
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 30 22 32 27 40 36 35 44 43 42 20 33 25 26 22 23
#> projected clustering = 28 19 30 27 39 39 40 44 31 47 25 33 23 25 27 23
#> ARI=0.638944841310744
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 28 19 30 27 39 39 40 44 31 47 25 33 23 25 27 23
#> projected clustering = 30 21 34 22 35 44 39 48 34 36 34 30 20 27 24 22
#> ARI=0.70469579869558
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 30 21 34 22 35 44 39 48 34 36 34 30 20 27 24 22
#> projected clustering = 23 24 34 20 39 42 44 51 33 37 30 30 16 28 29 20
#> ARI=0.710396638756063
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 23 24 34 20 39 42 44 51 33 37 30 30 16 28 29 20
#> projected clustering = 26 25 35 20 42 32 43 49 43 29 30 30 20 30 26 20
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 23 24 34 20 39 42 44 51 33 37 30 30 16 28 29 20
#> projected clustering = 20 21 37 24 38 43 38 46 41 27 31 34 17 29 32 22
#> ARI=0.740761014939628
#> probability matrix dimensions = 500 16
#> EPOCH: 10
#> current clustering = 20 21 37 24 38 43 38 46 41 27 31 34 17 29 32 22
#> projected clustering = 20 16 39 18 36 45 42 47 39 35 26 33 16 29 33 26
#> probability matrix dimensions = 500 16
#> EPOCH: 10
#> current clustering = 20 21 37 24 38 43 38 46 41 27 31 34 17 29 32 22
#> projected clustering = 26 15 34 17 34 36 43 54 37 35 28 30 27 29 33 22
#> ARI=0.747336672374071
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 26 15 34 17 34 36 43 54 37 35 28 30 27 29 33 22
#> projected clustering = 26 15 35 19 37 37 44 49 39 38 23 35 24 27 38 14
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 26 15 34 17 34 36 43 54 37 35 28 30 27 29 33 22
#> projected clustering = 22 19 34 14 41 32 39 55 34 40 27 34 25 29 31 24
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 26 15 34 17 34 36 43 54 37 35 28 30 27 29 33 22
#> projected clustering = 23 15 35 20 30 45 38 51 31 43 31 28 22 29 35 24
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 26 15 34 17 34 36 43 54 37 35 28 30 27 29 33 22
#> projected clustering = 26 16 41 16 23 43 35 55 34 29 33 35 24 35 30 25
#> probability matrix dimensions = 500 16
#> EPOCH: 11
#> current clustering = 26 15 34 17 34 36 43 54 37 35 28 30 27 29 33 22
#> projected clustering = 18 20 37 18 27 44 49 47 41 26 28 33 30 26 32 24
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.747336672374071.
#> ICP run: 2
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 243 257
#> ARI=0.795254009078026
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 243 257
#> projected clustering = 252 248
#> ARI=0.883366457789441
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 252 248
#> projected clustering = 246 254
#> ARI=0.936897581678391
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 246 254
#> projected clustering = 246 254
#> ARI=0.968192287637096
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 246 254
#> projected clustering = 245 255
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 246 254
#> projected clustering = 249 251
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 246 254
#> projected clustering = 247 253
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 246 254
#> projected clustering = 246 254
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 246 254
#> projected clustering = 242 258
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 127 127 123 123
#> projected clustering = 124 136 85 155
#> ARI=0.505655130498455
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 124 136 85 155
#> projected clustering = 138 128 96 138
#> ARI=0.683101815749058
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 138 128 96 138
#> projected clustering = 137 129 110 124
#> ARI=0.814566576111229
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 137 129 110 124
#> projected clustering = 140 129 114 117
#> ARI=0.821782487016266
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 140 129 114 117
#> projected clustering = 142 126 118 114
#> ARI=0.881342040005077
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 142 126 118 114
#> projected clustering = 139 126 124 111
#> ARI=0.931826790907787
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 139 126 124 111
#> projected clustering = 140 126 124 110
#> ARI=0.963230615890191
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 140 126 124 110
#> projected clustering = 141 127 132 100
#> probability matrix dimensions = 500 4
#> EPOCH: 8
#> current clustering = 140 126 124 110
#> projected clustering = 142 126 127 105
#> ARI=0.963588001669763
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 142 126 127 105
#> projected clustering = 139 126 130 105
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 142 126 127 105
#> projected clustering = 140 126 133 101
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 142 126 127 105
#> projected clustering = 142 126 126 106
#> probability matrix dimensions = 500 4
#> EPOCH: 9
#> current clustering = 142 126 127 105
#> projected clustering = 143 126 129 102
#> ARI=0.970245264019729
#> probability matrix dimensions = 500 4
#> EPOCH: 10
#> current clustering = 143 126 129 102
#> projected clustering = 141 126 131 102
#> ARI=0.970435510349295
#> probability matrix dimensions = 500 4
#> EPOCH: 11
#> current clustering = 141 126 131 102
#> projected clustering = 142 126 134 98
#> ARI=0.974273973234211
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 142 126 134 98
#> projected clustering = 137 126 133 104
#> probability matrix dimensions = 500 4
#> EPOCH: 12
#> current clustering = 142 126 134 98
#> projected clustering = 140 126 135 99
#> ARI=0.9891123089446
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 140 126 135 99
#> projected clustering = 130 126 134 110
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 140 126 135 99
#> projected clustering = 143 126 135 96
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 140 126 135 99
#> projected clustering = 141 126 134 99
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 140 126 135 99
#> projected clustering = 125 127 138 110
#> probability matrix dimensions = 500 4
#> EPOCH: 13
#> current clustering = 140 126 135 99
#> projected clustering = 142 127 131 100
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 70 70 50 49 63 63 68 67
#> projected clustering = 50 90 38 56 45 81 63 77
#> ARI=0.57268013864495
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 50 90 38 56 45 81 63 77
#> projected clustering = 51 93 44 50 35 90 56 81
#> ARI=0.799444109940251
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 51 93 44 50 35 90 56 81
#> projected clustering = 50 93 38 58 37 90 44 90
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 51 93 44 50 35 90 56 81
#> projected clustering = 52 93 48 48 34 91 77 57
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 51 93 44 50 35 90 56 81
#> projected clustering = 46 94 37 53 39 91 50 90
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 51 93 44 50 35 90 56 81
#> projected clustering = 52 96 43 46 39 85 64 75
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 51 93 44 50 35 90 56 81
#> projected clustering = 50 94 38 59 52 74 68 65
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 47 46 22 22 45 45 25 25 28 28 18 17 41 40 26 25
#> projected clustering = 24 67 17 27 54 27 33 29 15 32 25 20 41 39 23 27
#> ARI=0.464653531562265
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 24 67 17 27 54 27 33 29 15 32 25 20 41 39 23 27
#> projected clustering = 25 68 21 22 33 41 30 33 18 29 32 20 41 37 18 32
#> ARI=0.648681975631689
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 25 68 21 22 33 41 30 33 18 29 32 20 41 37 18 32
#> projected clustering = 24 71 21 21 20 45 22 39 7 53 43 18 32 36 16 32
#> ARI=0.686680138885083
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 24 71 21 21 20 45 22 39 7 53 43 18 32 36 16 32
#> projected clustering = 31 64 23 15 19 44 26 33 10 50 43 19 33 37 19 34
#> ARI=0.755485177036571
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 31 64 23 15 19 44 26 33 10 50 43 19 33 37 19 34
#> projected clustering = 26 66 21 18 19 44 33 26 14 50 48 16 25 38 19 37
#> ARI=0.768321186810094
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 26 66 21 18 19 44 33 26 14 50 48 16 25 38 19 37
#> projected clustering = 29 67 20 19 26 41 29 28 14 57 46 15 16 47 22 24
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 26 66 21 18 19 44 33 26 14 50 48 16 25 38 19 37
#> projected clustering = 29 66 13 17 24 36 37 27 16 51 54 14 27 35 21 33
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 26 66 21 18 19 44 33 26 14 50 48 16 25 38 19 37
#> projected clustering = 32 62 20 19 21 44 26 29 14 57 47 15 19 41 17 37
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 26 66 21 18 19 44 33 26 14 50 48 16 25 38 19 37
#> projected clustering = 30 64 20 19 14 42 32 28 20 39 57 14 36 34 24 27
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 26 66 21 18 19 44 33 26 14 50 48 16 25 38 19 37
#> projected clustering = 30 64 21 23 32 48 23 26 13 58 33 15 27 35 20 32
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.768321186810094.
#> ICP run: 3
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 255 245
#> ARI=0.838732948808876
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 255 245
#> projected clustering = 258 242
#> ARI=0.86836051504992
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 258 242
#> projected clustering = 251 249
#> ARI=0.944673239129708
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 251 249
#> projected clustering = 250 250
#> ARI=0.976096097153539
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 250 250
#> projected clustering = 261 239
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 250 250
#> projected clustering = 245 255
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 250 250
#> projected clustering = 253 247
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 250 250
#> projected clustering = 256 244
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 250 250
#> projected clustering = 252 248
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 125 125 125 125
#> projected clustering = 122 122 130 126
#> ARI=0.570082150811321
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 122 122 130 126
#> projected clustering = 116 127 125 132
#> ARI=0.743534931137085
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 116 127 125 132
#> projected clustering = 114 127 123 136
#> ARI=0.849602575314537
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 114 127 123 136
#> projected clustering = 112 126 122 140
#> ARI=0.905618944041145
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 112 126 122 140
#> projected clustering = 111 126 109 154
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 112 126 122 140
#> projected clustering = 115 126 124 135
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 112 126 122 140
#> projected clustering = 110 126 118 146
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 112 126 122 140
#> projected clustering = 104 127 107 162
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 112 126 122 140
#> projected clustering = 106 126 114 154
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 56 56 70 70 63 63 61 61
#> projected clustering = 46 61 64 87 45 81 43 73
#> ARI=0.54123484914936
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 46 61 64 87 45 81 43 73
#> projected clustering = 29 76 66 94 38 88 33 76
#> ARI=0.725561093762805
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 29 76 66 94 38 88 33 76
#> projected clustering = 34 69 67 84 46 81 42 77
#> ARI=0.774261769737241
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 34 69 67 84 46 81 42 77
#> projected clustering = 37 68 62 78 38 88 44 85
#> ARI=0.825857825784347
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 37 68 62 78 38 88 44 85
#> projected clustering = 32 71 55 79 40 86 49 88
#> ARI=0.899053273334009
#> probability matrix dimensions = 500 8
#> EPOCH: 6
#> current clustering = 32 71 55 79 40 86 49 88
#> projected clustering = 32 70 55 76 42 84 49 92
#> ARI=0.901752172178415
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 32 70 55 76 42 84 49 92
#> projected clustering = 30 71 50 81 44 82 46 96
#> ARI=0.923493565834838
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 30 71 50 81 44 82 46 96
#> projected clustering = 30 69 52 85 42 84 45 93
#> probability matrix dimensions = 500 8
#> EPOCH: 8
#> current clustering = 30 71 50 81 44 82 46 96
#> projected clustering = 31 71 46 81 42 84 50 95
#> ARI=0.936810537294432
#> probability matrix dimensions = 500 8
#> EPOCH: 9
#> current clustering = 31 71 46 81 42 84 50 95
#> projected clustering = 30 71 53 79 44 82 47 94
#> ARI=0.94160969369401
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 30 71 53 79 44 82 47 94
#> projected clustering = 31 69 54 80 46 80 43 97
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 30 71 53 79 44 82 47 94
#> projected clustering = 29 72 53 75 51 75 52 93
#> probability matrix dimensions = 500 8
#> EPOCH: 10
#> current clustering = 30 71 53 79 44 82 47 94
#> projected clustering = 31 68 54 80 44 82 48 93
#> ARI=0.944333317190972
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 31 68 54 80 44 82 48 93
#> projected clustering = 31 68 58 83 44 82 43 91
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 31 68 54 80 44 82 48 93
#> projected clustering = 32 66 55 81 46 80 49 91
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 31 68 54 80 44 82 48 93
#> projected clustering = 33 65 59 78 44 82 47 92
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 31 68 54 80 44 82 48 93
#> projected clustering = 31 69 54 79 43 83 47 94
#> probability matrix dimensions = 500 8
#> EPOCH: 11
#> current clustering = 31 68 54 80 44 82 48 93
#> projected clustering = 30 68 62 74 47 79 46 94
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 34 34 27 27 41 41 40 40 47 46 22 22 24 24 16 15
#> projected clustering = 24 43 16 39 33 51 31 52 39 54 15 27 18 26 13 19
#> ARI=0.514247211924704
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 24 43 16 39 33 51 31 52 39 54 15 27 18 26 13 19
#> projected clustering = 31 31 16 39 30 49 34 46 29 64 21 27 23 25 15 20
#> ARI=0.661179482085389
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 31 31 16 39 30 49 34 46 29 64 21 27 23 25 15 20
#> projected clustering = 32 31 16 35 27 52 32 53 50 41 22 26 17 32 12 22
#> ARI=0.681766195825721
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 32 31 16 35 27 52 32 53 50 41 22 26 17 32 12 22
#> projected clustering = 28 32 14 40 32 51 28 54 58 33 15 29 15 34 15 22
#> ARI=0.750208928470526
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 28 32 14 40 32 51 28 54 58 33 15 29 15 34 15 22
#> projected clustering = 29 30 12 42 29 51 35 52 42 42 13 34 15 35 16 23
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 28 32 14 40 32 51 28 54 58 33 15 29 15 34 15 22
#> projected clustering = 40 21 18 36 27 58 23 45 59 39 15 27 17 39 15 21
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 28 32 14 40 32 51 28 54 58 33 15 29 15 34 15 22
#> projected clustering = 35 24 20 39 26 53 33 49 61 29 14 33 16 29 14 25
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 28 32 14 40 32 51 28 54 58 33 15 29 15 34 15 22
#> projected clustering = 34 23 14 39 31 52 28 51 55 43 14 29 15 32 17 23
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 28 32 14 40 32 51 28 54 58 33 15 29 15 34 15 22
#> projected clustering = 32 25 16 41 26 58 32 51 28 57 12 30 16 35 18 23
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.750208928470526.
#> ICP run: 4
#> probability matrix dimensions = 500 2
#> EPOCH: 1
#> current clustering = 250 250
#> projected clustering = 265 235
#> ARI=0.83873398505553
#> probability matrix dimensions = 500 2
#> EPOCH: 2
#> current clustering = 265 235
#> projected clustering = 258 242
#> ARI=0.898501161140504
#> probability matrix dimensions = 500 2
#> EPOCH: 3
#> current clustering = 258 242
#> projected clustering = 256 244
#> ARI=0.936897707538242
#> probability matrix dimensions = 500 2
#> EPOCH: 4
#> current clustering = 256 244
#> projected clustering = 249 251
#> ARI=0.944673189756076
#> probability matrix dimensions = 500 2
#> EPOCH: 5
#> current clustering = 249 251
#> projected clustering = 248 252
#> ARI=0.97609610020095
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 248 252
#> projected clustering = 246 254
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 248 252
#> projected clustering = 252 248
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 248 252
#> projected clustering = 251 249
#> probability matrix dimensions = 500 2
#> EPOCH: 6
#> current clustering = 248 252
#> projected clustering = 245 255
#> ARI=0.976096118043035
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 245 255
#> projected clustering = 248 252
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 245 255
#> projected clustering = 241 259
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 245 255
#> projected clustering = 244 256
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 245 255
#> projected clustering = 248 252
#> probability matrix dimensions = 500 2
#> EPOCH: 7
#> current clustering = 245 255
#> projected clustering = 244 256
#> ARI=0.976096137703704
#> probability matrix dimensions = 500 2
#> EPOCH: 8
#> current clustering = 244 256
#> projected clustering = 248 252
#> probability matrix dimensions = 500 2
#> EPOCH: 8
#> current clustering = 244 256
#> projected clustering = 245 255
#> probability matrix dimensions = 500 2
#> EPOCH: 8
#> current clustering = 244 256
#> projected clustering = 248 252
#> probability matrix dimensions = 500 2
#> EPOCH: 8
#> current clustering = 244 256
#> projected clustering = 248 252
#> probability matrix dimensions = 500 2
#> EPOCH: 8
#> current clustering = 244 256
#> projected clustering = 248 252
#> probability matrix dimensions = 500 4
#> EPOCH: 1
#> current clustering = 128 128 122 122
#> projected clustering = 107 140 121 132
#> ARI=0.532065808397644
#> probability matrix dimensions = 500 4
#> EPOCH: 2
#> current clustering = 107 140 121 132
#> projected clustering = 110 131 140 119
#> ARI=0.672679552854405
#> probability matrix dimensions = 500 4
#> EPOCH: 3
#> current clustering = 110 131 140 119
#> projected clustering = 117 126 134 123
#> ARI=0.742739185419994
#> probability matrix dimensions = 500 4
#> EPOCH: 4
#> current clustering = 117 126 134 123
#> projected clustering = 114 127 134 125
#> ARI=0.821432387454403
#> probability matrix dimensions = 500 4
#> EPOCH: 5
#> current clustering = 114 127 134 125
#> projected clustering = 118 127 139 116
#> ARI=0.831106673271641
#> probability matrix dimensions = 500 4
#> EPOCH: 6
#> current clustering = 118 127 139 116
#> projected clustering = 115 127 150 108
#> ARI=0.921458939150972
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 115 127 150 108
#> projected clustering = 107 128 159 106
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 115 127 150 108
#> projected clustering = 114 127 156 103
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 115 127 150 108
#> projected clustering = 111 126 160 103
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 115 127 150 108
#> projected clustering = 104 127 163 106
#> probability matrix dimensions = 500 4
#> EPOCH: 7
#> current clustering = 115 127 150 108
#> projected clustering = 130 127 130 113
#> probability matrix dimensions = 500 8
#> EPOCH: 1
#> current clustering = 58 57 54 54 64 63 75 75
#> projected clustering = 35 73 50 58 59 68 47 110
#> ARI=0.571917744633526
#> probability matrix dimensions = 500 8
#> EPOCH: 2
#> current clustering = 35 73 50 58 59 68 47 110
#> projected clustering = 33 68 40 69 39 88 43 120
#> ARI=0.7591367000387
#> probability matrix dimensions = 500 8
#> EPOCH: 3
#> current clustering = 33 68 40 69 39 88 43 120
#> projected clustering = 35 62 41 69 33 94 45 121
#> ARI=0.845979383640427
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 35 62 41 69 33 94 45 121
#> projected clustering = 41 51 47 68 34 95 45 119
#> probability matrix dimensions = 500 8
#> EPOCH: 4
#> current clustering = 35 62 41 69 33 94 45 121
#> projected clustering = 42 55 41 72 30 96 42 122
#> ARI=0.905821117045133
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 42 55 41 72 30 96 42 122
#> projected clustering = 41 57 40 74 33 93 42 120
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 42 55 41 72 30 96 42 122
#> projected clustering = 37 58 40 74 32 95 42 122
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 42 55 41 72 30 96 42 122
#> projected clustering = 43 55 36 69 31 95 42 129
#> probability matrix dimensions = 500 8
#> EPOCH: 5
#> current clustering = 42 55 41 72 30 96 42 122
#> projected clustering = 44 53 42 73 32 94 44 118
#> ARI=0.9173783968805
#> probability matrix dimensions = 500 8
#> EPOCH: 6
#> current clustering = 44 53 42 73 32 94 44 118
#> projected clustering = 43 55 42 74 34 92 41 119
#> probability matrix dimensions = 500 8
#> EPOCH: 6
#> current clustering = 44 53 42 73 32 94 44 118
#> projected clustering = 38 60 48 75 40 86 42 111
#> probability matrix dimensions = 500 8
#> EPOCH: 6
#> current clustering = 44 53 42 73 32 94 44 118
#> projected clustering = 40 59 39 73 33 93 45 118
#> ARI=0.919429669540258
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 40 59 39 73 33 93 45 118
#> projected clustering = 37 62 39 74 43 83 45 117
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 40 59 39 73 33 93 45 118
#> projected clustering = 43 55 41 76 38 88 44 115
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 40 59 39 73 33 93 45 118
#> projected clustering = 38 60 46 76 31 93 41 115
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 40 59 39 73 33 93 45 118
#> projected clustering = 42 51 42 71 35 95 47 117
#> probability matrix dimensions = 500 8
#> EPOCH: 7
#> current clustering = 40 59 39 73 33 93 45 118
#> projected clustering = 41 58 45 77 35 91 42 111
#> probability matrix dimensions = 500 16
#> EPOCH: 1
#> current clustering = 20 20 47 46 20 19 59 59 30 29 17 16 37 36 23 22
#> projected clustering = 16 24 43 45 13 26 37 83 27 33 19 18 26 43 22 25
#> ARI=0.525606720175476
#> probability matrix dimensions = 500 16
#> EPOCH: 2
#> current clustering = 16 24 43 45 13 26 37 83 27 33 19 18 26 43 22 25
#> projected clustering = 19 17 48 39 13 27 26 94 26 33 22 18 48 26 16 28
#> ARI=0.655272225926455
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 19 17 48 39 13 27 26 94 26 33 22 18 48 26 16 28
#> projected clustering = 19 28 36 50 16 25 49 73 15 34 25 15 37 29 18 31
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 19 17 48 39 13 27 26 94 26 33 22 18 48 26 16 28
#> projected clustering = 22 24 29 57 12 28 40 80 16 35 27 12 33 40 18 27
#> probability matrix dimensions = 500 16
#> EPOCH: 3
#> current clustering = 19 17 48 39 13 27 26 94 26 33 22 18 48 26 16 28
#> projected clustering = 18 25 49 41 13 26 27 93 22 32 24 13 48 23 19 27
#> ARI=0.666878575178302
#> probability matrix dimensions = 500 16
#> EPOCH: 4
#> current clustering = 18 25 49 41 13 26 27 93 22 32 24 13 48 23 19 27
#> projected clustering = 16 24 46 34 15 26 32 91 16 41 34 13 44 23 19 26
#> ARI=0.719017455319634
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 16 24 46 34 15 26 32 91 16 41 34 13 44 23 19 26
#> projected clustering = 13 32 38 33 22 23 40 83 14 38 43 13 29 36 16 27
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 16 24 46 34 15 26 32 91 16 41 34 13 44 23 19 26
#> projected clustering = 16 20 42 34 34 21 28 83 22 41 37 13 42 22 19 26
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 16 24 46 34 15 26 32 91 16 41 34 13 44 23 19 26
#> projected clustering = 13 21 36 31 16 23 35 95 20 45 46 12 41 20 28 18
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 16 24 46 34 15 26 32 91 16 41 34 13 44 23 19 26
#> projected clustering = 16 23 44 31 19 26 32 87 18 40 40 13 24 41 26 20
#> probability matrix dimensions = 500 16
#> EPOCH: 5
#> current clustering = 16 24 46 34 15 26 32 91 16 41 34 13 44 23 19 26
#> projected clustering = 18 24 29 37 30 23 20 89 25 31 48 12 42 25 24 23
#> ARI=0.725082283085276
#> probability matrix dimensions = 500 16
#> EPOCH: 6
#> current clustering = 18 24 29 37 30 23 20 89 25 31 48 12 42 25 24 23
#> projected clustering = 21 20 31 38 29 24 18 93 26 31 45 12 28 38 22 24
#> ARI=0.781909113382413
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 21 20 31 38 29 24 18 93 26 31 45 12 28 38 22 24
#> projected clustering = 27 21 28 39 27 25 18 93 19 31 47 12 35 29 26 23
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 21 20 31 38 29 24 18 93 26 31 45 12 28 38 22 24
#> projected clustering = 29 20 20 45 27 21 33 92 21 28 48 12 26 33 21 24
#> probability matrix dimensions = 500 16
#> EPOCH: 7
#> current clustering = 21 20 31 38 29 24 18 93 26 31 45 12 28 38 22 24
#> projected clustering = 17 26 29 39 33 25 21 89 23 31 45 12 25 39 23 23
#> ARI=0.79146881232699
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 17 26 29 39 33 25 21 89 23 31 45 12 25 39 23 23
#> projected clustering = 21 24 29 42 38 22 23 80 25 29 42 12 35 33 19 26
#> probability matrix dimensions = 500 16
#> EPOCH: 8
#> current clustering = 17 26 29 39 33 25 21 89 23 31 45 12 25 39 23 23
#> projected clustering = 27 14 28 46 36 26 17 90 22 34 39 12 29 35 19 26
#> ARI=0.808159013434938
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 27 14 28 46 36 26 17 90 22 34 39 12 29 35 19 26
#> projected clustering = 30 16 26 47 24 29 27 89 26 26 41 12 35 29 19 24
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 27 14 28 46 36 26 17 90 22 34 39 12 29 35 19 26
#> projected clustering = 20 19 24 50 31 27 29 81 21 39 40 12 36 25 25 21
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 27 14 28 46 36 26 17 90 22 34 39 12 29 35 19 26
#> projected clustering = 28 15 32 42 30 29 21 86 21 35 39 12 32 33 25 20
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 27 14 28 46 36 26 17 90 22 34 39 12 29 35 19 26
#> projected clustering = 31 19 28 48 24 29 26 93 18 30 37 12 34 27 26 18
#> probability matrix dimensions = 500 16
#> EPOCH: 9
#> current clustering = 27 14 28 46 36 26 17 90 22 34 39 12 29 35 19 26
#> projected clustering = 36 15 24 54 41 24 35 71 19 28 35 12 25 37 21 23
#> ICP converged at EPOCH 1.
#> Maximum ARI reached: 0.808159013434938.
#> 
#> Divisive ICP clustering completed successfully.
#> 
#> Predicting cell cluster probabilities using ICP models...
#> Prediction of cell cluster probabilities completed successfully.
#> 
#> Multi-level integration completed successfully.

# 2) Compute reference PCA & UMAP
ref <- RunPCA(ref, return.model = TRUE, pca.method = "stats")
#> Divisive ICP: selecting ICP tables multiple of 4
set.seed(123)
ref <- RunUMAP(ref, return.model = TRUE)

# 3) Project & predict query cell labels 
set.seed(1024)
map <- ReferenceMapping(ref = ref, query = query, ref.label = "cell_type", scale.query.by = NULL, project.umap = TRUE)

# Calculate accuracy
preds_x_truth <- table(map$coral_labels, map$cell_type)
preds_x_truth
#>                 
#>                  aDC B mem B naive CD4 mem CD4 naive CD8 eff CD8 T HSC
#>   aDC             13     0       0       0         0       0     0   0
#>   B mem            0    37      16       0         0       1     0   2
#>   B naive          0     7      68       0         0       0     0   0
#>   CD4 mem          0     0       0      83        26       0     2   0
#>   CD4 naive        0     0       0      36       160       0    21   1
#>   CD8 eff          0     0       0       3         0      99     7   0
#>   CD8 T            0     1       1       5        10       7    52   0
#>   HSC              0     0       0       0         0       0     0   2
#>   Megakaryocyte    0     0       0       0         0       0     0   0
#>   Monocyte         2     0       0       0         0       0     0   0
#>   CD16+ monocyte   0     0       0       0         0       0     0   0
#>   NK               0     0       0       0         0       1     0   0
#>   pDC              0     0       0       0         0       0     0   0
#>   Treg             0     0       0       1         0       0     0   0
#>                 
#>                  Megakaryocyte Monocyte CD16+ monocyte  NK pDC Treg
#>   aDC                        0        6              0   0   0    0
#>   B mem                      0        0              0   0   0    0
#>   B naive                    0        0              0   0   2    0
#>   CD4 mem                    0        0              0   0   0    5
#>   CD4 naive                  0        0              0   0   0   10
#>   CD8 eff                    1        0              0   2   0    0
#>   CD8 T                      1        0              0   0   0    0
#>   HSC                        0        0              0   0   0    0
#>   Megakaryocyte              0        0              0   0   0    0
#>   Monocyte                   3      172             12   1   0    0
#>   CD16+ monocyte             1        0             64   0   0    0
#>   NK                         0        0              0  54   0    0
#>   pDC                        0        0              0   0   2    0
#>   Treg                       0        0              0   0   0    0
sum(diag(preds_x_truth)) / sum(preds_x_truth) # 0.806 : 81%
#> [1] 0.806