Divisive Iterative Clustering Projection (ICP) clustering

The function implements divisive Iterative Clustering Projection (ICP) clustering: a supervised learning-based clustering, which maximizes clustering similarity between the clustering and its projection by logistic regression, doing it in a divisive clustering manner.

Usage

RunDivisiveICP(
  normalized.data = NULL,
  batch.label = NULL,
  k = 16,
  d = 0.3,
  r = 5,
  C = 5,
  reg.type = "L1",
  max.iter = 200,
  icp.batch.size = Inf,
  train.with.bnn = TRUE,
  train.k.nn = 10,
  train.k.nn.prop = NULL,
  cluster.seed = NULL,
  divisive.method = "random",
  allow.free.k = FALSE,
  ari.cutoff = 0.5
)

Arguments

normalized.data

A sparse matrix (dgCMatrix) containing normalized feature expression data with cells in rows and features in columns. Default is NULL.

batch.label

A character vector with batch labels corresponding to the cells given in normalized.data. The character batch labels need to be named with the cells names given in the rows of normalized.data. By default NULL, i.e., cells are sampled evenly regardless their batch.

k

A positive integer power of two, i.e., 2**n, where n>0, specifying the number of clusters in the last Iterative Clustering Projection (ICP) round. Decreasing k leads to smaller cell populations diversity and vice versa. Default is 16, i.e., the divisive clustering 2 -> 4 -> 8 -> 16 is performed.

d

A numeric that defines how many cells per cluster should be down- and oversampled (d in ceiling(N/k*d)), when stratified.downsampling=FALSE, or what fraction should be downsampled in the stratified approach ,stratified.downsampling=TRUE. Default is 0.3.

r

A positive integer that denotes the number of reiterations performed until the algorithm stops. Default is 5.

C

Cost of constraints violation (C) for L1-regulatization. Default is 0.3.

reg.type

"L1" for LASSO and "L2" for Ridge. Default is "L1".

max.iter

A positive integer that denotes the maximum number of iterations performed until the algorithm ends. Default is 200.

icp.batch.size

A positive integer that specifies how many cells to randomly select for each ICP run from the complete data set. This is a new feature intended to speed up the process with larger data sets. Default is Inf, which means using all cells.

train.with.bnn

Train data with batch nearest neighbors. Default is TRUE. Only used if batch.label is given.

train.k.nn

Train data with batch nearest neighbors using k nearest neighbors. Default is 10. Only used if train.with.bnn is TRUE.

train.k.nn.prop

A numeric (higher than 0 and lower than 1) corresponding to the fraction of cells per cluster to use as train.k.nn nearest neighbors. Default is NULL meaning that the number of train.k.nn nearest neighbors is equal to train.k.nn. If given, train.k.nn parameter is ignored and train.k.nn is calculated based on train.k.nn.prop. A vector with different proportions for the different divisive clustering rounds can be given, otherwise the same value is given for all.

cluster.seed

A cluster seed to start and guide the clustering to more reproducible clusterings across runs (factor). Default is NULL. Otherwise, a random clustering takes place to start divisive clustering with ICP.

divisive.method

Divisive method (character). One of "random" (randomly sample two clusters out of every cluster previously found), "cluster" or "cluster.batch" (sample two clusters out of every cluster previously found based on the cluster probability distribution across batches or per batch). By default "random".

allow.free.k

Allow free k (logical). Allow ICP algorithm to decrease the k given in case it does not find k target clusters. By default FALSE.

ari.cutoff

Include ICP models and probability tables with an Adjusted Rand Index higher than ari.cutoff (numeric). By default 0.5. A value that can range between 0 (include all) and lower than 1.

Value

A list that includes the probability matrix and the clustering similarity measures: ARI, NMI, etc.