Sample cells based on principal components distribution

Samples cells based on their distributions along one principal component

Usage

SamplePCACells(
  data,
  batch = NULL,
  q.split = 0.5,
  p = 30,
  use.pc = "PC1",
  center = TRUE,
  scale. = TRUE
)

Arguments

data

Data to compute PCA and sample cells from. Rows and columns should represent cells and features, respectively.

batch

Batch cell label identity (character) matching cells giving in data. Use NULL in the absence of batches. If the batch is given the cells are sampled in a batch wise manner, otherwise the cells are sampled without any grouping factor. By default is NULL.

q.split

Split (cell) batch principal component distribution by this quantile (numeric). By default 0.5, i.e., median.

p

Number of principal components to compute (integer). By default 30.

use.pc

Which principal component should be used for sampling cells per batch. By default "PC1", i.e., first principal component is used.

center

Should the features given in data centered before performing the PCA (logical). By default TRUE.

scale.

Should the features given in data scaled before performing the PCA (logical). By default TRUE.

Value

A factor with cell cluster identities (two clusters).